+++ projects/tweakium-core/src/main/kotlin/site/siredvin/tweakium/modules/peripheral/api/IPeripheralPlugin.kt
@@ -21,4 +21,5 @@ interface IPeripheralPlugin {
 
     fun onFirstAttach() {}
     fun onLastDetach() {}
+    fun collectConfiguration(data: MutableMap<String, Any>) {}
 }