+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2025-09-14
+
+### Fixed
+
+- `inventory` now support item slot limit correctly
+- Add configuration information for plugins, exposing limits for `item_storage`, `fluid_storage
+ and add inventoryAPIVersion for `inventory`
 ## [1.0.0] - 2025-03-17
 
 ### Added